Fix race condition on entering STBY
The MCU latches STBY with E, which means it can take a bit over one E cycle to
fully enter standby. The current decoder code tries to drive AS immediately
after it pulls STBY low - and it can't move it vs the MCU. Programming on the
first command only works if 'AS' just so happens to be at the right level
already.

This code makes the first WR strobe (after a 250ms cooldown) affect the STBY pin
only. We need a tiny bit of extra code to enter STBY now, but on the plus side
this should also allow implementing a 'reset' command withing driving the bus in
any way.