ASM: functions to play/stop a given MIDI note on a given channel
The main trick here are lookup tables for FNUM/BLOCK to avoid having to calculate what these would be on the fly. The tables are tuned so MIDI 69 = A4 = 440Hz comes out correctly when using the 13MHz crystal that's on my breadboard right now. This will need fixing when using an up-to-spec 14.32 MHz crystal.