Fix reset issue for the MIDI board
There are some reset issues with the TL16C2550. This change makes it so that the
RESET line is held high for a few ms after power-on before going low. That seems
to get rid of some weirdness. There is some speculation online that the chip
needs to see a low-to-high transition to properly reset - that might be a future
improvement if this ends up being flaky. None of this is documented in the
datasheet.