MIDI Synth: Improved MIDI velocity to OPL3 attenuation translation algorithm
It's a piecewise linear approximation of 40*log(V/127) with cutoffs at MIDI
velocities 32 and 64. This strikes a balance between easy to compute and
sounding close to MIDI spec. It's a bit too loud for low volumes. MIDI decays
faster here. An additional step at velocity 16 could help if this needs
improving.