Emulator: increase preformance by avoiding some expensive locking
A good 25% of emulator time is spent just locking locks on each tick to find out
whether we have interrupts or whether there is serial data waiting. These
conditions are almost always false and we can drastically improve performance by
replacing them with an atomic flag that is much cheaper to test in the false
case.