Emulator: improve performance by avoiding redundant lookups
Before this we looked up each instruction 4 times: twice each in tick() and
execute(). In each of these we did a contains() to figure out if the instruction
is valid, followed by a lookup of the actual instruction. Replacing them with
one find() and some const references gets rid of 3 lookups and two Instruction
creation/destructions too.