Emulator: improve performance by removing instruction map
The instruction lookups in std::map made up ~35% of the cost of tick(). This
removes the map in favor of an std::array. The implementation now looks a bit
wonky as we still initialize a map just to copy it all over into an array. To be
fixed at some point in the future.