Emulator: Make the IOPort API more explicit and more complete
The current API makes a mess of read/write and input/output. After this rename
it should be clearer how one has to interact with a port.

The current API also doesn't offer a way to drive input pin changes nor register
interest in those input changes. This is now addressed too.